+++ byers/src/main.rs
@@ -1,4 +1,5 @@
 use std::net::ToSocketAddrs;
+use chrono::Utc;
 
 use commands::help::help;
 use fred::{
@@ -7,6 +8,7 @@ use fred::{
     prelude::{ClientLike, PubsubInterface, RedisClient},
     types::{PerformanceConfig, ReconnectPolicy, RedisConfig, RedisValue},
 };
+use poise::FrameworkError;
 use poise::serenity_prelude::Activity;
 use sqlx::postgres::PgPoolOptions;
 use tokio::task::JoinSet;
@@ -27,6 +29,7 @@ use crate::{
         songs::song,
         version::version,
         youtube::youtube,
+        listen
     },
     communication::ByersUnixStream,
     db::DbSong,
@@ -64,6 +67,7 @@ async fn main() {
         user(),
         minigames::command(),
         add(),
+        listen(),
     ];
 
     info!("Loading {} commands...", commands.len());
@@ -202,6 +206,29 @@ async fn main() {
                     }
 
                     Ok(())
+                })
+            },
+            on_error: |error| {
+                Box::pin(async move {
+                    match error {
+                        FrameworkError::CooldownHit { remaining_cooldown, ctx } => {
+                            ctx.send(|m| {
+                                m.embed(|e| {
+                                    e.title("You are too fast!")
+                                        .description(format!(
+                                            "You can use that command again {}.",
+                                            (Utc::now().naive_utc() + chrono::Duration::from_std(remaining_cooldown).unwrap()).relative_time()
+                                        ))
+                                }).ephemeral(true)
+                            }).await.expect_or_log("unable to send cooldown message");
+                        },
+                        _ => {
+                            let result = poise::builtins::on_error(error).await;
+                            if let Err(error) = result {
+                                error!("Discord error: {}", error);
+                            }
+                        }
+                    }
                 })
             },
             ..Default::default()